Add files via upload
THIS IS THE ANSWER TO THE QUESTION NUMBER 23 OF NLTK BOOK.

THIS IS THE QUESTION:
Write a recursive function lookup(trie, key) that looks up a key in a trie, and returns the
value it finds. Extend the function to return a word when it is uniquely determined by its
prefix (e.g. vanguard is the only word that starts with vang-, so lookup(trie, 'vang') should
return the same thing as lookup(trie, 'vanguard')).